% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_file_size.R
\name{get_one_file_size}
\alias{get_one_file_size}
\title{get_one_file_size Title}
\usage{
get_one_file_size(path_fichier)
}
\arguments{
\item{path_fichier}{le chemin vers le fichier}
}
\value{
df un dataframe avec les infos sur le fichier
}
\description{
get_one_file_size Title
}
\examples{
get_one_file_size(glue::glue("{system.file(package = 'toolboxMael')}/nyc_squirrels_sample.csv"))
}
